% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_paths.R
\name{example_paths}
\alias{example_paths}
\title{Paths to Examples}
\usage{
example_paths()
}
\value{
Named sequence of characters. Each character string points
at an example file in the package.
}
\description{
Paths to Examples
}
