% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams.mylearn-package.R
\docType{package}
\name{exams.mylearn-package}
\alias{exams.mylearn-package}
\alias{exams.mylearn}
\title{Question Generation in the 'MyLearn' XML Format}
\description{
Randomized multiple-select and single-select
question generation for the 'MyLearn' platform. Question templates
in the form of the R/exams package
are transformed into XML format required by 'MyLearn'.
}
