% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams.wuvienna-package.R
\docType{package}
\name{exams.wuvienna-package}
\alias{exams.wuvienna-package}
\alias{exams.wuvienna}
\title{Question Generation in the 'MyLearn' XML Format}
\description{
Randomized multiple-select and single-select
question generation for the 'MyLearn' platform at the Vienna
University of Economics and Business. Question templates
in the form of the R/exams package
are transformed into XML format required by 'MyLearn'.
}
