% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2mylearn.R
\name{exams2mylearn}
\alias{exams2mylearn}
\title{Exam Generation for the MyLearn Platform}
\usage{
exams2mylearn(filename, name, n, dir = ".")
}
\arguments{
\item{filename}{(character) absolute or relative path to the exercise template.
Usually simply a filename pointing at a .Rmd file in the working directory}

\item{name}{(character) unique name of output files}

\item{n}{(integer) number of random variants to create}

\item{dir}{(character, optional) output directory, defaults to the currect working directory}
}
\value{
As a side effect, the function produces a zip file in the working
directory. The exact path to the zip file is returned invisibly.
}
\description{
The Vienna University of Economics and Business has a special
XML format on its teaching platform MyLearn. \code{exams2mylearn} transforms
input files in the R/exams to XML files and zips them. The resulting
zip file can be directly uploaded to the MyLearn platform after having
contected the MyLearn development team.
}
\note{
The development team has to turn on the upload functionality on
a per course basis.
}
\examples{
ex_files <- example_paths()
exams2mylearn(ex_files["mixture"], "various_options", 2)
\dontrun{
exams2mylearn(ex_files["plot"], "boxplot_exercise", 40)
exams2mylearn(ex_files["single_choice"], "single_choice_exercise", 40)
}
}
